% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candid.R
\name{candid}
\alias{candid}
\alias{candid_data}
\alias{candid_lookup}
\alias{candid_metadata}
\alias{candid_data_ids}
\alias{candid_lookup_ids}
\title{CANDID Datasets}
\usage{
candid_data(dataset_id)

candid_lookup(lookup_id)

candid_metadata()

candid_data_ids()

candid_lookup_ids()
}
\arguments{
\item{dataset_id}{ID for a particular CANDID dataset. Run
\code{candid_data_ids()} for the list of available options.}

\item{lookup_id}{ID for a particular CANDID lookup table. Run
\code{candid_lookup_ids()} for the list of available options.}
}
\description{
Access datasets used in the paper describing the
Canadian Disease Incidence Dataset (CANDID).
}
\section{Functions}{
\itemize{
\item \code{candid_data()}: Return a data frame of one of the datasets used in the
CANDID paper.

\item \code{candid_lookup()}: Return a data frame of one of the datasets used in the
CANDID paper.

\item \code{candid_metadata()}: Return a list of lists, each of which represents
the metadata for a CANDID dataset.

\item \code{candid_data_ids()}: Return the IDs of all the datasets
used in the CANDID paper.

\item \code{candid_lookup_ids()}: Return the IDs of all the lookup tables
used in the CANDID paper.

}}
