% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featured.R
\name{featured}
\alias{featured}
\alias{featured_data}
\alias{featured_metadata}
\alias{featured_ids}
\title{Featured Datasets}
\usage{
featured_data(dataset_id, ...)

featured_metadata()

featured_ids()
}
\arguments{
\item{dataset_id}{ID for a particular featured dataset. Run
\code{featured_ids()} for the list of available options.}

\item{...}{Character vectors for filtering the data on specific columns,
with one vector for each column. See examples below for the syntax for
different types of columns. The following list gives instructions for
columns that might be available for your dataset. If they are not available
you will get a message telling you what columns are in the dataset
that you are asking for.
\itemize{
\item location : Location descriptor as reported in the original source.
\item iso_3166 : ISO-3166 codes for locations corresponding to countries.
\item iso_3166_2 : ISO-3166-2 codes for locations corresponding to sub-national jurisdictions.
\item date : Date of time point at which a value is reported, in ISO-8601 format (i.e. YYYY-MM-DD). Must be in the form \<start date\>..<end date\>.
\item period_start_date : The first date within the time interval for which a count was reported, in ISO-8601 format (i.e. YYYY-MM-DD). Must be in the form \<start date\>..<end date\>.
\item period_end_date : The last date within the time interval for which a count was reported, in ISO-8601 format (i.e. YYYY-MM-DD). Must be in the form \<start date\>..<end date\>.
\item period_mid_date : Date in the (rounded down) middle of a time interval for which a count was reported, in ISO-8601 format (i.e. YYYY-MM-DD). Must be in the form \<start date\>..<end date\>.
\item historical_disease_family : Disease family as reported by the original source.
\item historical_disease : Disease name as reported by the original source.
\item disease : Disease name harmonized over data sources.
\item nesting_disease : Disease name that encapsulates this disease in a hierarchy of diseases.
\item basal_disease : Base disease name that encapsulates this disease in a hierarchy of diseases.
\item icd_9 : Disease codes for version 9 of the International Statistical Classification of Diseases and Related Health Problems, published by the World Health Organization.
\item icd_7 : Disease codes for version 7 of the International Statistical Classification of Diseases and Related Health Problems, published by the World Health Organization.
\item historical_disease_subclass : Disease subclass as reported by the original source -- cf. Disease
\item icd_9_subclass : Disease codes associated with a subclass identified by the original source. Codes are for version 9 of the International Statistical Classification of Diseases and Related Health Problems, published by the World Health Organization.
\item icd_7_subclass : Disease codes associated with a subclass identified by the original source. Codes are for version 7 of the International Statistical Classification of Diseases and Related Health Problems, published by the World Health Organization.
\item lower_age : Lower age bound in years of all individuals associated with a particular record. The first item must either be a number interval of the form \<min\>-\<max\> or 'none' (meaning no filter is applied to the case numbers). Additional items are meant to be any 'unavailable values' like 'Not available', 'Not reportable', or 'null'.
\item upper_age : Upper age bound in years of all individuals associated with a particular record. The first item must either be a number interval of the form \<min\>-\<max\> or 'none' (meaning no filter is applied to the case numbers). Additional items are meant to be any 'unavailable values' like 'Not available', 'Not reportable', or 'null'.
\item sex : Sex of all individuals associated with a particular record as reported by the original source.
\item cases_this_period : Number of cases reported for this period, at the end of this period -- as reported by the original source. Includes different types of missing values as reported by the original source. Type "string" allows different types of missing values to be specified as reported by the original source and allows users to choose which type of missing values to work with. The first item must either be a number interval of the form \<min\>-\<max\> or 'none' (meaning no filter is applied to the case numbers). Additional items are meant to be any 'unavailable values' like 'Not available', 'Not reportable', or 'null'.
\item cases_prev_period : Number of cases reported for the previous period, at the end of this current period -- as reported by the original source. Includes different types of missing values as reported by the original source. Type "string" allows different types of missing values to be specified as reported by the original source and allows users to choose which type of missing values to work with. The first item must either be a number interval of the form \<min\>-\<max\> or 'none' (meaning no filter is applied to the case numbers). Additional items are meant to be any 'unavailable values' like 'Not available', 'Not reportable', or 'null'.
\item cases_cum_report_year : Total number of cases up until this current period, in the current year -- as reported by the original source. Includes different types of missing values as reported by the original source. Type "string" allows different types of missing values to be specified as reported by the original source and allows users to choose which type of missing values to work with. The first item must either be a number interval of the form \<min\>-\<max\> or 'none' (meaning no filter is applied to the case numbers). Additional items are meant to be any 'unavailable values' like 'Not available', 'Not reportable', or 'null'.
\item cases_cum_prev_year : Total number of cases up until this current period, in the previous year -- as reported by the original source. Includes different types of missing values as reported by the original source. Type "string" allows different types of missing values to be specified as reported by the original source and allows users to choose which type of missing values to work with.
\item cases_median_prev_5_years : Median number of weekly cases in this current period, with the median taken over the previous five years -- as reported by the original source. Includes different types of missing values as reported by the original source. Type "string" allows different types of missing values to be specified as reported by the original source and allows users to choose which type of missing values to work with. The first item must either be a number interval of the form \<min\>-\<max\> or 'none' (meaning no filter is applied to the case numbers). Additional items are meant to be any 'unavailable values' like 'Not available', 'Not reportable', or 'null'.
\item cases_cum_median_prev_5_years : Median number of cumulative cases by this current period, with the median taken over the previous five years -- as reported by the original source. Includes different types of missing values as reported by the original source. Type "string" allows different types of missing values to be specified as reported by the original source and allows users to choose which type of missing values to work with. The first item must either be a number interval of the form \<min\>-\<max\> or 'none' (meaning no filter is applied to the case numbers). Additional items are meant to be any 'unavailable values' like 'Not available', 'Not reportable', or 'null'.
\item population : Estimated total number of individuals associated with a particular record. The first item must either be a number interval of the form \<min\>-\<max\> or 'none' (meaning no filter is applied to the case numbers). Additional items are meant to be any 'unavailable values' like 'Not available', 'Not reportable', or 'null'.
\item cause : Cause of mortality as reported by the original source.
\item location_type : Location type of recorded data (continents, countries, canadian-provinces, canadian-cities, ontario-cities, uk-cities, london-parishes) for a given record.
\item time_scale : Time scale of recorded data (wk, 2wk, mo, qr, yr) for a given record. This field is a qualitative and approximate measure of the time scale over which data were counted. For example, monthly (mo) data might include four-weekly or even five-weekly data. See columns like days_this_period for a quantitative description of the time scale if it is available, or produce your own description using period_start_date and period_end_date.
}}
}
\description{
Access featured datasets and associated metadata. These datasets required
substantial preparation to enhance their utility for research and analysis.
}
\section{Functions}{
\itemize{
\item \code{featured_data()}: Return a data frame of a featured
dataset, possibly filtered.

\item \code{featured_metadata()}: Return a list of lists, each of which represents
the metadata for a featured dataset.

\item \code{featured_ids()}: Return the IDs of all the featured
datasets in the repository.

}}
\examples{
featured_ids()
atlantic_polio_1950s = featured_data("canmod-cdi-normalized"
  , iso_3166_2 = c("CA-NL", "CA-NS", "CA-PE", "CA-NB")
  , basal_disease = "poliomyelitis"
  , period_end_date = "1950-01-01..1959-12-31"
)

}
