% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\docType{data}
\name{interactive}
\alias{interactive}
\alias{docs_url}
\alias{docs_url_staging}
\alias{docs_url_local}
\title{Links to IIDDA API Interactive Documentation}
\format{
An object of class \code{character} of length 0.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
docs_url

docs_url_staging

docs_url_local
}
\description{
Links to IIDDA API Interactive Documentation
}
\section{Functions}{
\itemize{
\item \code{docs_url}: Link to interactive documentation.

\item \code{docs_url_staging}: Link to interactive documentation for a staging
environment.

\item \code{docs_url_local}: Localhost link to interactive documentation for
a development environment, if it exists.

}}
\keyword{datasets}
