% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{urls}
\alias{urls}
\alias{url_scans}
\alias{url_digitizations}
\alias{url_affected_scripts}
\title{Data Provenance}
\usage{
url_scans(
  scan_ids,
  dataset_ids = character(),
  metadata = ops_staging$metadata(dataset_ids = dataset_ids)
)

url_digitizations(
  digitization_ids,
  dataset_ids = character(),
  metadata = ops_staging$metadata(dataset_ids = dataset_ids)
)

url_affected_scripts(
  digitization_id,
  dataset_ids = character(),
  metadata = ops_staging$metadata(dataset_ids = dataset_ids)
)
}
\arguments{
\item{scan_ids}{Character vector of identifiers for scans.}

\item{dataset_ids}{Identifiers for datasets that will
restrict the search for dependencies to metadata for
these datasets.}

\item{metadata}{List of IIDDA DataCite metadata over
a collection of datasets.}

\item{digitization_ids}{Character vector of identifiers for digitizations.}

\item{digitization_id}{Single digitization identifier.}
}
\description{
Get URLs to dependencies of datasets for investigating
data provenance. See \code{vignette("Provenance")} for an
illustration. See
\href{https://github.com/canmod/iidda-staging#identifiers}{iidda-staging}
for information on IIDDA identifiers. If you do not have access
to this link, please contact the
\href{https://github.com/canmod/iidda-tools/tree/main/R/iidda.api#maintainer}{maintainer}.
}
\section{Functions}{
\itemize{
\item \code{url_scans()}: Convert scan identifiers into URLs that link
to associated scans (typically PDF files).

\item \code{url_digitizations()}: Convert digitization identifiers into URLs that link
to associated digitizations (typically Excel files).

\item \code{url_affected_scripts()}: Returns URLs to all prep scripts that could have
their output modified if the identified digitization were modified.

}}
