% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\docType{package}
\name{iidda.api-package}
\alias{iidda.api}
\alias{iidda.api-package}
\title{\pkg{iidda.api}}
\description{
IIDDA API.
}
\details{
IIDDA is the International Infectious Disease Data Archive.
This archive has an \href{https://math.mcmaster.ca/iidda/api/docs}{API}
(Application Programming Interface)
for accessing data and potentially for building applications.
This R package provides a simple wrapper to this API so that
the datasets are returned as data frames.
\subsection{Listing the Datasets}{

\if{html}{\out{<div class="sourceCode {r "dataset-names"}">}}\preformatted{iidda.api::ops_staging$metadata() |> names()
}\if{html}{\out{</div>}}
}

\subsection{Getting Datasets}{

Just choose one of the identifiers above, and pass it to the
\code{raw_csv} function in the API.

\if{html}{\out{<div class="sourceCode {r, eval = FALSE}">}}\preformatted{cdi_1975 = iidda.api::ops_staging$raw_csv(
   dataset_ids = c(
     "cdi_ca_1975_wk_prov_statcan",
     "cdi_ca_1976_wk_prov_statcan"
   )
)
}\if{html}{\out{</div>}}

You can pass more than one dataset.

\if{html}{\out{<div class="sourceCode {r, eval = FALSE}">}}\preformatted{cdi_1975 = iidda.api::ops_staging$raw_csv(
   dataset_ids = "cdi_ca_1975_wk_prov_statcan"
)
}\if{html}{\out{</div>}}
}

\subsection{All Metadata by Dataset}{

\if{html}{\out{<div class="sourceCode {r, eval = FALSE}">}}\preformatted{iidda.api::ops_staging$metadata()
}\if{html}{\out{</div>}}

On my browser at least, it is easier to look at in JSON form rather than
R list form. The JSON metadata can be directly accessed using
the URL form of the API:
\url{https://math.mcmaster.ca/iidda/api/metadata}.
}

\subsection{CANMOD Digitization Project}{

The communicable disease incidence (CDI) data collected as part of
the \href{https://canmod.net/digitization}{CANMOD digitization project}
can be accessed using \code{resource_type = "CANMOD CDI"} with the \code{filter}
function.

\if{html}{\out{<div class="sourceCode {r, eval = FALSE}">}}\preformatted{canmod_cdi = iidda.api::ops_staging$filter(
     resource_type = "CANMOD CDI"
   , iso_3166 = "CA" ## country code for canada
)
}\if{html}{\out{</div>}}
}
}
\author{
\strong{Maintainer}: Steve Walker \email{swalk@mcmaster.ca}

}
