% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\docType{package}
\name{iidda.api-package}
\alias{iidda.api}
\alias{iidda.api-package}
\title{\pkg{iidda.api}}
\description{
IIDDA API.
}
\details{
IIDDA is the International Infectious Disease Data Archive.
This archive has an \href{https://math.mcmaster.ca/iidda/api/docs}{API}
(Application Programming Interface)
for accessing data and potentially for building applications.
This R package provides a simple wrapper to this API so that
the datasets are returned as data frames.

Useful links for people who just want to get data.
\itemize{
\item \code{?featured}
\item \code{vignette("QuickStart")}
\item \code{vignette("Provenance")}
}

More advanced users might be interested in the lower-level wrapper
of the API operations here: \code{?ops}.
}
\author{
\strong{Maintainer}: Steve Walker \email{swalk@mcmaster.ca}

}
