% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\docType{package}
\name{iidda.api-package}
\alias{iidda.api}
\alias{iidda.api-package}
\title{\pkg{iidda.api}}
\description{
R binding to the IIDDA API.
}
\author{
\strong{Maintainer}: First Last \email{swalk@mcmaster.ca}

}
