% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\docType{package}
\name{iidda.api-package}
\alias{iidda.api}
\alias{iidda.api-package}
\title{\pkg{iidda.api}}
\description{
IIDDA API.
}
\details{
IIDDA is the International Infectious Disease Data Archive.
This archive has an \href{https://math.mcmaster.ca/iidda/api/docs}{API}
(Application Programming Interface)
for accessing data and potentially for building applications.
This R package provides a simple wrapper to this API so that
the datasets are returned as data frames.
\subsection{Listing the Datasets}{

There are many datasets in the archive, but the most important ones are
called compilations. These compilations generally pull together data
from several sources and are designed to provide convenience for users
who just want to get the data and start using it right away. To list
the IDs of the compilation datasets one may use the following R command.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{featured_ids()
#> [1] "canmod-cdi-harmonized"   "canmod-cdi-normalized"  
#> [3] "canmod-cdi-unharmonized" "canmod-mort-harmonized" 
#> [5] "canmod-mort-normalized"  "canmod-pop-normalized"
}\if{html}{\out{</div>}}

To get the list of all of the datasets in the archive one may use the
following.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ops_staging$metadata() |> names()
#>   [1] "cdi_wc_ca_1924-55_wk_prov"                    
#>   [2] "cdi_ca_1993-2001_quart_prov_hc"               
#>   [3] "cdi_ca_1990-1992_mn_prov_hc"                  
#>   [4] "cdi_ca_1984-89_4wk_prov_statcan"              
#>   [5] "cdi_ca_1979-83_4wk_prov_statcan"              
#>   [6] "cdi_ca_1972_wk_prov"                          
#>   [7] "cdi_ca_1971_wk_prov"                          
#>   [8] "cdi_ca_1970_wk_prov"                          
#>   [9] "cdi_ca_1969_wk_prov"                          
#>  [10] "cdi_ca_1968_wk_prov"                          
#>  [11] "cdi_ca_1967_wk_prov_mcgill"                   
#>  [12] "cdi_ca_1966_wk_prov_mcgill"                   
#>  [13] "cdi_ca_1965_wk_prov_mcgill"                   
#>  [14] "cdi_ca_1964_wk_prov_mcgill"                   
#>  [15] "canmod-cdi-normalized"                        
#>  [16] "cdi_sk_1910-1927_mn"                          
#>  [17] "mort_on_1940-1947_mn"                         
#>  [18] "mort_on_1903-1939_mn"                         
#>  [19] "cdi_wc_ca_1924_wk_prov"                       
#>  [20] "cdi_on_1940-1947_mn"                          
#>  [21] "cdi_on_1903-1939_mn"                          
#>  [22] "cdi_ca_1962_wk_prov_dbs"                      
#>  [23] "cdi_ab_2004-19_wk_age"                        
#>  [24] "canmod-cdi-unharmonized"                      
#>  [25] "phac-reporting-schedule"                      
#>  [26] "cdi_on_1939-1989_wk"                          
#>  [27] "canmod-time-scale-cross-check"                
#>  [28] "canmod-location-cross-check"                  
#>  [29] "canmod-disease-cross-check"                   
#>  [30] "canmod-cdi-harmonized"                        
#>  [31] "cdi_qc_1915-1925_mn_county_municipality"      
#>  [32] "cdi_ca_1973_wk_prov"                          
#>  [33] "cdi_ca_1960_wk_prov_dbs"                      
#>  [34] "prototype-dataset"                            
#>  [35] "cdi_yf_ca_1945-1955_wk_prov"                  
#>  [36] "cdi_granuing_ca_1946-55_wk_prov"              
#>  [37] "canmod-pop-normalized"                        
#>  [38] "pop_ca_1921-71_an_age_prov_sex"               
#>  [39] "phac-cdi-portal"                              
#>  [40] "mort_ca_2000-2020_wk_prov_causes"             
#>  [41] "mort_ca_1990-1999_wk_prov_causes"             
#>  [42] "mort_ca_1980-1989_wk_prov_causes"             
#>  [43] "mort_ca_1970-1979_wk_prov_causes"             
#>  [44] "mort_ca_1960-1969_wk_prov_causes"             
#>  [45] "mort_ca_1950-1959_wk_prov_causes"             
#>  [46] "canmod-mort-normalized"                       
#>  [47] "pop_ca_1971-2021_an_age_prov_sex"             
#>  [48] "pop_ca_1871-1921_10-yearly_prov_age_sex"      
#>  [49] "mort_plag_uk_1661-1688_wk_par"                
#>  [50] "cdi_sk_1910-1927_mn_harmonized"               
#>  [51] "cdi_ca_1963_wk_prov_dbs"                      
#>  [52] "canmod-mort-harmonized"                       
#>  [53] "cdi_ca_1924-55_wk_prov_dbs_statcan_harmonized"
#>  [54] "cdi_on_1990-2021_wk_harmonized"               
#>  [55] "cdi_ca_1979-89_4wk_prov_harmonized"           
#>  [56] "cdi_tball_ca_1924-55_wk_prov"                 
#>  [57] "cdi_chanc_ca_1924-55_wk_prov"                 
#>  [58] "cdi_ca_1990-2001_quart_prov_harmonized"       
#>  [59] "cdi_ca_1975-78_wk_prov_harmonized"            
#>  [60] "cdi_ca_1961_wk_prov_dbs"                      
#>  [61] "cdi_ca_1959_wk_prov_dbs_w4_52"                
#>  [62] "cdi_ca_1959_wk_prov_dbs_w1_3"                 
#>  [63] "cdi_ca_1958_wk_prov_dbs"                      
#>  [64] "cdi_ca_1957_wk_prov_dbs"                      
#>  [65] "cdi_ca_1956_wk_prov_dbs"                      
#>  [66] "cdi_bot_ca_1933-55_wk_prov"                   
#>  [67] "cdi_ab_2004-19_wk_harmonized"                 
#>  [68] "cdi_uf_ca_1928-55_wk_prov"                    
#>  [69] "cdi_typhpara_ca_1952-55_wk_prov"              
#>  [70] "cdi_typh_ca_1924-52_wk_prov"                  
#>  [71] "cdi_tbunspec_ca_1933-55_wk_prov"              
#>  [72] "cdi_tbpulm_ca_1933-55_wk_prov"                
#>  [73] "cdi_tbmisc_ca_1933-55_wk_prov"                
#>  [74] "cdi_syphunspec_ca_1944-55_wk_prov"            
#>  [75] "cdi_syphtertother_ca_1944-55_wk_prov"         
#>  [76] "cdi_syphtertneuro_ca_1944-55_wk_prov"         
#>  [77] "cdi_syphtertcard_ca_1944-55_wk_prov"          
#>  [78] "cdi_syphsec_ca_1944-55_wk_prov"               
#>  [79] "cdi_syphprim_ca_1944-55_wk_prov"              
#>  [80] "cdi_syphother_ca_1944-51_wk_prov"             
#>  [81] "cdi_syphlatent_ca_1944-55_wk_prov"            
#>  [82] "cdi_syphcong_ca_1944-55_wk_prov"              
#>  [83] "cdi_syphall_ca_1924-55_wk_prov"               
#>  [84] "cdi_sst_ca_1924-55_wk_prov"                   
#>  [85] "cdi_smallpox_ca_1924-55_wk_prov"              
#>  [86] "cdi_sf_ca_1924-55_wk_prov"                    
#>  [87] "cdi_salm_ca_1944-55_wk_prov"                  
#>  [88] "cdi_rub_ca_1924-55_wk_prov"                   
#>  [89] "cdi_poliopar_ca_1949-55_wk_prov"              
#>  [90] "cdi_polio_ca_1924-55_wk_prov"                 
#>  [91] "cdi_pneu_ca_1924-55_wk_prov"                  
#>  [92] "cdi_paratyph_ca_1924-55_wk_prov"              
#>  [93] "cdi_optneo_ca_1944-55_wk_prov"                
#>  [94] "cdi_on_1939-1989_wk_harmonized"               
#>  [95] "cdi_nspecurethritis_ca_1949-55_wk_prov"       
#>  [96] "cdi_mumps_ca_1924-55_wk_prov"                 
#>  [97] "cdi_meningosep_ca_1952-54_wk_prov"            
#>  [98] "cdi_meningivar_ca_1925-55_wk_prov"            
#>  [99] "cdi_meningisep_ca_1935-55_wk_prov"            
#> [100] "cdi_malar_ca_1929-55_wk_prov"                 
#> [101] "cdi_lymphgran_ca_1946-55_wk_prov"             
#> [102] "cdi_jauncat_ca_1948-55_wk_prov"               
#> [103] "cdi_jaun_ca_1927-55_wk_prov"                  
#> [104] "cdi_impet_ca_1924-55_wk_prov"                 
#> [105] "cdi_hep_ca_1945-55_wk_prov"                   
#> [106] "cdi_erys_ca_1924-55_wk_prov"                  
#> [107] "cdi_enceph_ca_1924-34_wk_prov"                
#> [108] "cdi_dysunspec_ca_1927-55_wk_prov"             
#> [109] "cdi_dysbac_ca_1924-55_wk_prov"                
#> [110] "cdi_dysam_ca_1927-55_wk_prov"                 
#> [111] "cdi_dysall_ca_1929-55_wk_prov"                
#> [112] "cdi_diphth_ca_1924-55_wk_prov"                
#> [113] "cdi_diarent_ca_1945-55_wk_prov"               
#> [114] "cdi_diar_ca_1937-44_wk_prov"                  
#> [115] "cdi_cp_ca_1924-55_wk_prov"                    
#> [116] "cdi_ca_1968-72_wk_prov_harmonized"            
#> [117] "cdi_ca_1964-67_wk_prov_harmonized"            
#> [118] "cdi_tbnpulm_ca_1935-55_wk_prov"               
#> [119] "cdi_qc_1915-1925_mn_harmonized"               
#> [120] "cdi_poliounspec_ca_1949-55_wk_prov"           
#> [121] "cdi_polionpar_ca_1949-55_wk_prov"             
#> [122] "cdi_otherVD_ca_1944-55_wk_prov"               
#> [123] "cdi_on_1940-1947_mn_harmonized"               
#> [124] "cdi_on_1903-1939_mn_harmonized"               
#> [125] "cdi_gonall_ca_1924-55_wk_prov"                
#> [126] "cdi_ca_1956-63_1973-74_wk_prov_harmonized"    
#> [127] "cdi_on_1990-2021_wk"                          
#> [128] "cdi_ca_1978_wk_prov_statcan"                  
#> [129] "cdi_ca_1977_wk_prov_statcan"                  
#> [130] "cdi_ca_1976_wk_prov_statcan"                  
#> [131] "cdi_ca_1975_wk_prov_statcan"                  
#> [132] "pop_ca_1871-2021_harmonized"                  
#> [133] "mort_uk_1842-1950"                            
#> [134] "mort_uk_1661-1845"                            
#> [135] "mort_uk_1642-1845_wk_davenport"               
#> [136] "london-mort-harmonized"                       
#> [137] "cdi_man_2017jan-jun_mn"                       
#> [138] "cdi_man_2016_mn"                              
#> [139] "cdi_man_2015_mn"                              
#> [140] "cdi_man_2014_mn"                              
#> [141] "cdi_man_2013_mn"                              
#> [142] "cdi_man_2012_mn"                              
#> [143] "cdi_man_2011_mn"                              
#> [144] "cdi_man_2010_mn"                              
#> [145] "cdi_man_2009_mn"                              
#> [146] "cdi_man_2008_mn"                              
#> [147] "cdi_man_2007_mn"                              
#> [148] "cdi_man_2006_mn"                              
#> [149] "cdi_man_2005_mn"                              
#> [150] "cdi_man_2004_mn"                              
#> [151] "cdi_man_2004-17_mn_harmonized"                
#> [152] "cdi_ca_2001-2006_qr_prov_ccdr_harmonized"     
#> [153] "cdi_ca_2001-2006_qr_prov_ccdr"                
#> [154] "cdi_ca_1974_wk_prov"                          
#> [155] "acm_uk_1842-1930_age"                         
#> [156] "acm_uk_1661-1845"
}\if{html}{\out{</div>}}
}

\subsection{Getting Datasets}{

To read one of these datasets into an R data frame, choose one of the
identifiers above and pass it to the \code{raw_csv} operation.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{canmod_cdi = iidda.api::ops_staging$raw_csv(
   dataset_ids = "canmod-cdi-normalized"
)
#> Now converting some fields from character to numeric or date. 
#> You can turn this off with
#> options(iidda_api_all_char = TRUE). 
#> Do not display any iidda.api messages with
#> options(iidda_api_msgs = FALSE).
#> Messages displayed at most once per R session.
#> Now sorting by date. 
#> You can turn this off with
#> options(iidda_api_date_sort = FALSE). 
#> Do not display any iidda.api messages with
#> options(iidda_api_msgs = FALSE).
#> Messages displayed at most once per R session.
print(canmod_cdi)
#> # A tibble: 935,603 x 21
#>    basal_disease      cases_this_period days_this_period digitization_id disease
#>    <chr>                          <dbl>            <dbl> <chr>           <chr>  
#>  1 smallpox                         196               31 cdi_mort_on_19~ smallp~
#>  2 scarlet-fever                    500               31 cdi_mort_on_19~ scarle~
#>  3 diphtheria                       510               31 cdi_mort_on_19~ diphth~
#>  4 measles                            6               31 cdi_mort_on_19~ measles
#>  5 whooping-cough                     9               31 cdi_mort_on_19~ whoopi~
#>  6 typhoid-or-malaria                65               31 cdi_mort_on_19~ typhoi~
#>  7 tuberculosis                     176               31 cdi_mort_on_19~ tuberc~
#>  8 smallpox                         153               28 cdi_mort_on_19~ smallp~
#>  9 scarlet-fever                    550               28 cdi_mort_on_19~ scarle~
#> 10 diphtheria                       200               28 cdi_mort_on_19~ diphth~
#> # i 935,593 more rows
#> # i 16 more variables: historical_disease <chr>,
#> #   historical_disease_family <chr>, historical_disease_subclass <chr>,
#> #   iso_3166 <chr>, iso_3166_2 <chr>, location <chr>, location_type <chr>,
#> #   nesting_disease <chr>, original_dataset_id <chr>, period_end_date <date>,
#> #   period_mid_date <date>, period_start_date <date>, population <dbl>,
#> #   record_origin <chr>, scan_id <chr>, time_scale <chr>
}\if{html}{\out{</div>}}
}

\subsection{All Metadata by Dataset}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{iidda.api::ops_staging$metadata()
}\if{html}{\out{</div>}}

On my browser at least, it is easier to look at in JSON form rather than
R list form. The JSON metadata can be directly accessed using
the URL form of the API:
\url{https://math.mcmaster.ca/iidda/api/metadata}.
}
}
\author{
\strong{Maintainer}: Steve Walker \email{swalk@mcmaster.ca}

}
