% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindings.R
\docType{data}
\name{ops}
\alias{ops}
\alias{ops_local}
\alias{ops_staging}
\title{IIDDA API Operations}
\format{
An object of class \code{try-error} of length 1.

An object of class \code{try-error} of length 1.

An object of class \code{list} of length 6.
}
\usage{
ops

ops_local

ops_staging
}
\description{
Objects containing the functions associated with API functions
documented \href{https://math.mcmaster.ca/iidda/api/docs}{here}.
}
\section{Functions}{
\itemize{
\item \code{ops}: List containing available operations from the IIDDA API
as \code{R} functions

\item \code{ops_local}: Operations list for a local development environment,
if it exists

\item \code{ops_staging}: Operations list for a staging environment, if it exists

}}
\examples{
## Print out the available functions.
names(ops_staging)

## Access functions with a dollar sign. For example, this command
## will give weekly incidence data in PEI in January of 1940.
ops_staging$filter(
     resource_type = "CANMOD CDI"
   , iso_3166_2 = "CA-PE"
   , period_end_date = "1940-01-01..1940-02-01"
   , time_scale = "wk"
)

}
\keyword{datasets}
